+++ CHANGELOG.md
@@ -67,6 +67,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed grid within GridSpec test [#806](https://github.com/ie3-institute/simona/issues/806)
 - Fixed log entry after power flow calculation [#814](https://github.com/ie3-institute/simona/issues/814)
 - Not stopping correctly on failed power flow if configured  to stop [#800](https://github.com/ie3-institute/simona/issues/800)
+- Simulation stopping at unhandled messages in `DBFSAlgorithm` [#821](https://github.com/ie3-institute/simona/issues/821)
 
 ## [3.0.0] - 2023-08-07